Prepare keras for making ResourceVariables as CompositeTensors.
We are going to let ResourceVariable be a subclass of CompositeTensor. Changes in this CL are necessary to not break existing code.

Specifically, to track resource variables embedded in composite tensors, we will need to manually expand composite tensors layer by layer instead of replying on tf.nest.

Currently resource variables are atoms and considered to have the same structure as tensors. So we could have one branch to be a resource variable and the other branch to be a tensor. After making resource variable as composite tensors, resource variables will be tf.nest sequences instead of atoms. To avoid the type spec mismatch, we replace resource variables with tf.nest atoms just for the purpose of tf.nest.assert_same_structure.

PiperOrigin-RevId: 456609757